+++ lib/app_profiler.rb
@@ -30,6 +30,7 @@ module AppProfiler
   require "app_profiler/request_parameters"
   require "app_profiler/profiler"
   require "app_profiler/profile"
+  require "app_profiler/server"
 
   mattr_accessor :logger, default: Logger.new($stdout)
   mattr_accessor :root
@@ -45,6 +46,7 @@ module AppProfiler
   mattr_accessor :storage, default: Storage::FileStorage
   mattr_accessor :viewer, default: Viewer::SpeedscopeViewer
   mattr_accessor :middleware, default: Middleware
+  mattr_accessor :server, default: Server
 
   class << self
     def run(*args, &block)